feat: comprehensive codebase refactoring and organization
This commit addresses a comprehensive refactoring epic, improving code organization, naming, and architectural patterns.

Key changes include:
- Decoupled server logic by introducing a service registry and standardized service interfaces.
- Centralized the Spotify API client to ensure a singleton pattern and moved common API response parsing to a utility file.
- Consolidated volume state management by renaming `useVolumePreference` to `useVolume` and updating its usages.
- Improved code organization and naming by renaming `getHrZoneProps` to `calculateWorkoutMetrics` and updating its usages.
- Updated import statements to use TypeScript path aliases consistently.
- Fixed all linting errors.

The planned refactoring of `apiService.ts` was not performed as the file does not exist in the current codebase.
