+++ app/client/spotify-selection/page.tsx
@@ -16,8 +16,9 @@ import Typography from '@mui/material/Typography'
 import dynamic from 'next/dynamic'
 import { useEffect, useState } from 'react'
 import VolumeControl from '../../../components/Spotify/VolumeControl' // I will recreate this temporarily
-import useVolume from '@/hooks/useVolume'
-import { SpotifyCommandMessage } from '@/types/websocket'
+import useVolumePreference from '../../../hooks/useVolumePreference'
+import { useWebSocket } from '@/context/WebSocketContext'
+import { SpotifyCommandMessage } from '../../../types/websocket'
 import { API_SPOTIFY_DEVICES } from '@/constants/apiEndpoints'
 
 const PlaylistSelector = dynamic(
@@ -73,7 +74,7 @@ const SpotifySelectionPage = () => {
       setSelectedDeviceId('')
     }
   }, [availableDevices, selectedDeviceId])
-  const { volume, setVolume } = useVolume()
+  const { volume, setVolume } = useVolumePreference()
 
   const handlePlaylistSelected = (uri: string) => {
     setSelectedPlaylistUri(uri)
