+++ hooks/useTimerSounds.ts
@@ -5,12 +5,12 @@
  */
 import { useCallback, useEffect, useRef } from 'react'
 import { audioManager } from '../utils/audioManager'
-import useVolume from './useVolume'
+import useVolumePreference from './useVolumePreference'
 import { useWebSocket } from '@/context/WebSocketContext'
 
 export const useTimerSounds = (isMuted: boolean = false) => {
   const { timerData } = useWebSocket()
-  const { volume } = useVolume()
+  const { volume } = useVolumePreference()
   const lastSoundEventId = useRef<number>(0)
 
   // Update audio volume when volume preference changes
