+++ services/spotifyPolling.ts
@@ -2,7 +2,15 @@ import { AccessToken, SpotifyApi, Device } from '@spotify/web-api-ts-sdk'
 import { ServerMessage, SpotifyData, SpotifyDevice } from '../types/websocket'
 import { SpotifyTokenManager } from './spotifyTokenManager.js'
 import logger from '../utils/logger.js'
-import { safeParseJSON } from '../utils/apiResponseUtils.js'
+
+// Utility: Safely parse JSON, fallback to text
+function safeParseJSON(input: string): unknown {
+  try {
+    return JSON.parse(input)
+  } catch {
+    return input // Return raw text if not JSON
+  }
+}
 
 // API endpoint constants (mostly managed by SDK now)
 // TOKEN_URL is handled by TokenManager or SDK
